Fix Tests and examples
