+++ examples/02_AABB_collisions.html
@@ -8,22 +8,23 @@
 <body></body>
 <script>
     const Engine = end3in.Engine,
-        Body = end3in.Body,
+        Rectangle = end3in.Rectangle,
         Collision = end3in.Collision,
         Graphics = end3in.Graphics,
-        Vec2 = end3in.Vector;
-    const engine = Engine.create();
-    const rect1 = Body.rect(0.5, 0.5, 3, 4, {'isStatic': true, 'color': "#141ae6EE", "rot": 0.22});
-    const rect2 = Body.rect(1.0, 1.5, 5, 6, {'isStatic': true, 'color': "#141ae6EE", "rot": 0.1});
+        Vec2 = end3in.Vec2;
+    const engine = new Engine();
+    const rect1 = new Rectangle(0.5, 0.5, 3, 4, 1, true, 0.22);
+    const rect2 = new Rectangle(1.0, 1.5, 5, 6, 1, true,  Math.PI / 5);
     engine.addBody(rect1);
     engine.addBody(rect2);
+    engine.g = 0
     engine.start();
 
-    let graphics = Graphics.create(600, 600, engine, true, 60, Vec2.create(0,0))
+    let graphics = new Graphics(600, 600, engine, true, 60, new Vec2(0,0))
     graphics.start()
 
     document.onmousemove = function (event) {
-        rect1.pos = graphics.clientToWorldPos(Vec2.create(event.clientX, event.clientY));
+        rect1.pos = graphics.clientToWorldPos(new Vec2(event.clientX, event.clientY));
         if (Collision.areColliding(rect1, rect2)) {
             rect1.color = "#e62131EE";
         } else {
