+++ examples/04_shapes.html
@@ -8,44 +8,25 @@
 <body></body>
 <script>
     const Engine = end3in.Engine,
-        Body = end3in.Body,
+        Rectangle = end3in.Rectangle,
         Collision = end3in.Collision,
         Graphics = end3in.Graphics,
-        Vec2 = end3in.Vector;
+        Vec2 = end3in.Vec2,
+        Polygon = end3in.Polygon;
 
-    let engine = Engine.create(debug = true, solver = "none");
-    const triangle = Body.polygon([ [0, 0.6], [1, -0.255],[-1, -0.255]], 1, 2, {
-        'isStatic': true,
-        'color': "#141ae6EE",
-        "rot": 0
-    }); // Math.PI / 4
-    const polygon1 = Body.polygon([ [1.2, 0.8],[0.69, -0.76], [0, -0.5],[-0.25, 0] ],
-        0.8, 4.2, {
-            'isStatic': true,
-            'color': "#141ae6EE",
-            "rot": Math.PI / 6
-        });
-    const polygon2 = Body.polygon([[-0.4, -0.20], [0.44, 0.37], [0.65, 0.25],[0.62, -0.40] ],
-        3.3, 2.0, {
-            'isStatic': true,
-            'color': "#141ae6EE",
-            "rot": Math.PI / 6
-        });
-    const rect1 = Body.rect(1, 1, 2, -1, {
-        'isStatic': true,
-        'color': "#141ae6EE",
-        "rot": Math.PI / 4
-    });
+    let engine = new Engine(false, "none");
+    const triangle = new Polygon([[0, 0.6], [1, -0.255], [-1, -0.255]], 1, 2, 1, true, 0);
+    const polygon1 = new Polygon([[1.2, 0.8], [0.69, -0.76], [0, -0.5], [-0.25, 0]], 0.8, 4.2, 1, true, Math.PI / 6);
+    const polygon2 = new Polygon([[-0.4, -0.20], [0.44, 0.37], [0.65, 0.25], [0.62, -0.40]], 3.3, 2.0, 1, true, Math.PI / 6);
+    const rect1 = new Rectangle(1, 1, 2, -1, 1, true, Math.PI / 4);
 
-
-    let graphics = Graphics.create(1000, 1000, engine, true)
+    let graphics = new Graphics(1000, 1000, engine, true)
 
     engine.addBody(triangle);
     engine.addBody(polygon1);
     engine.addBody(polygon2);
     engine.addBody(rect1);
     engine.start();
-
 
     document.addEventListener("keydown", event => {
         if (event.isComposing || event.code === 'ArrowLeft') {
@@ -58,7 +39,7 @@
     document.onmousemove = function (event) {
         graphics.clear()
         graphics.drawBodies()
-        triangle.pos = graphics.clientToWorldPos(Vec2.create(event.clientX, event.clientY))
+        triangle.pos = graphics.clientToWorldPos(new Vec2(event.clientX, event.clientY))
         triangle.color = "#141ae6EE";
 
         for (let gameObj of engine.gameObjects) {
