+++ examples/08_inside.html
@@ -8,31 +8,24 @@
 <body></body>
 <script>
     const Engine = end3in.Engine,
-        Body = end3in.Body,
+        Rectangle = end3in.Rectangle,
+        Polygon = end3in.Polygon,
         Collision = end3in.Collision,
         Graphics = end3in.Graphics,
-        Vec2 = end3in.Vector;
-    let engine = Engine.create();
-    const rect1 = Body.polygon([[-1, -0.255], [1, -0.255], [0, 0.6]], 1, 2, {
-        'isStatic': true,
-        'color': "#141ae6EE",
-        "rot": 0
-    }); // Math.PI / 4
-    const polygon2 = Body.rect(2, 2,
-        3.3, 2.0, {
-            'isStatic': true,
-            'color': "#141ae6EE",
-            "rot": Math.PI / 6
-        });
+        Vec2 = end3in.Vec2;
+    let engine = new Engine();
+    const rect1 = new Polygon([[-1, -0.255], [1, -0.255], [0, 0.6]], 1, 2, 1, true);
+    const polygon2 = new Rectangle(2, 2,
+        3.3, 2.0, 1,  true, Math.PI / 6);
     engine.addBody(rect1);
     engine.addBody(polygon2);
     engine.start();
 
-    const graphics = Graphics.create(1000, 1000, engine)
+    const graphics = new Graphics(1000, 1000, engine)
     graphics.start()
 
     document.onmousemove = function (event) {
-        const mousePos = graphics.clientToWorldPos(Vec2.create(event.clientX, event.clientY));
+        const mousePos = graphics.clientToWorldPos(new Vec2(event.clientX, event.clientY));
         rect1.pos = mousePos
         if (polygon2.isInside(mousePos)) {
             polygon2.color = "#e62131EE";
