+++ examples/12_transpose_solver.html
@@ -8,48 +8,40 @@
 <body></body>
 <script>
     const Engine = end3in.Engine,
-        Body = end3in.Body,
+        Rectangle = end3in.Rectangle,
         Collision = end3in.Collision,
         Graphics = end3in.Graphics,
-        Vec2 = end3in.Vector;
+        Polygon = end3in.Polygon,
+        Vec2 = end3in.Vec2;
 
-    let engine = Engine.create(false, 'translation');
+    let engine = new Engine(false, 'translation');
 
-    const obj = Body.polygon([[-0.5, 1.0], [2.0, 1.0], [0.5, 0.0]], 0, 0)
+    const obj = new Polygon([[-0.5, 1.0], [2.0, 1.0], [0.5, 0.0]], 0, 0)
     engine.addBody(obj)
 
     for (let i = 0; i < 10; i++) {
-        const isStatic = Math.random() >= 0.5
-        let options = {
-            isStatic: false
-        }
-        if (isStatic) {
-            options["color"] = "#000"
-        }
-        engine.addBody(Body.rect(0.5, 0.5, (Math.random() - 0.5) * 5, (Math.random() - 0.5) * 5, options));
-        engine.addBody(Body.polygon([[0.0, 0.0], [0.5, 0.0], [0.3, -0.8]], (Math.random() - 0.5) * 5, (Math.random() - 0.5) * 5));
+        engine.addBody(new Rectangle(0.5, 0.5, (Math.random() - 0.5) * 5, (Math.random() - 0.5) * 5, 1, Math.random() >= 0.8));
+        engine.addBody(new Polygon([[0.0, 0.0], [0.5, 0.0], [0.3, -0.8]], (Math.random() - 0.5) * 5, (Math.random() - 0.5) * 5, 1, Math.random() >= 0.8));
     }
-    engine.addBody(Body.rect(7,1, -0, 4, {isStatic: true}));
+    engine.addBody(new Rectangle(7, 1, -0, 4, 1, true));
 
-    engine.g = -1;
     engine.start();
 
-
-    const graphics = Graphics.create(1500, 1500, engine, {debug: true}, scale = 60, cameraPos = Vec2.create(6,6))
+    const graphics = new Graphics(1500, 1500, engine, {debug: true}, scale = 60, cameraPos = new Vec2(6, 6))
     graphics.start()
 
     document.addEventListener("keydown", event => {
         if (event.isComposing || event.code === 'ArrowUp') {
-            obj.applyImpulse(Vec2.create(0, 50))
+            obj.applyImpulse(new Vec2(0, 50))
         }
         if (event.isComposing || event.code === 'ArrowDown') {
-            obj.applyImpulse(Vec2.create(0, -50))
+            obj.applyImpulse(new Vec2(0, -50))
         }
         if (event.isComposing || event.code === 'ArrowLeft') {
-            obj.applyImpulse(Vec2.create(50, 0))
+            obj.applyImpulse(new Vec2(50, 0))
         }
         if (event.isComposing || event.code === 'ArrowRight') {
-            obj.applyImpulse(Vec2.create(-50,0 ))
+            obj.applyImpulse(new Vec2(-50, 0))
         }
     });
 </script>
