+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "end3in",
-  "version": "1.0.0",
+  "version": "2.0.0",
   "description": "Simple physic engine ",
   "main": "build/end3in.js",
   "scripts": {
@@ -24,7 +24,7 @@
   },
   "homepage": "https://github.com/mrugacz95/end3in#readme",
   "devDependencies": {
-    "jest": "^29.1.2",
+    "ts-jest": "^29.0.3",
     "ts-loader": "^9.4.1",
     "typescript": "^4.8.4",
     "webpack": "^5.74.0",
